Replace page /index with /posts. Update navigation. Add jsconfig.json
